+++ ctwin32/__init__.py
@@ -22,7 +22,7 @@
 #
 ################################################################################
 
-version = "1.8.0"
+version = "1.9.0"
 
 from types import SimpleNamespace as _namespace
 from .wtypes import *